Fix behaviour of scopes `rootScope` has been renamed to `globalScope`.
First, scopes should not recursively find items from parents — there
should only be a global scope and a local scope when within a function.
If a thing cannot be found in the local scope, find it in the global
scope instead of the parent scope.

Also fixed setting scope items since the parent scope was being searched
first before using the local one. This now uses the same searching
method as getting scope items.